Created a new QuotingService and split out most of the functionality from QuoteModule to it. Also fixed an issue with manual quoting and beautified the code
