+++ Orabot/Modules/QuoteModule.cs
@@ -7,12 +7,19 @@ using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
 using Orabot.Objects;
+using Orabot.Services;
 
 namespace Orabot.Modules
 {
 	public class QuoteModule : ModuleBase<SocketCommandContext>
 	{
+		private readonly QuotingService _quotingService;
 		private readonly string[] _trustedRoles = ConfigurationManager.AppSettings["TrustedRoles"].Split(';').Select(x => x.Trim()).ToArray();
+
+		public QuoteModule(QuotingService quotingService)
+		{
+			_quotingService = quotingService;
+		}
 
 		[Command("quote")]
 		[Summary("Quotes an arbitrary message with optional author, source channel and timestamp. Limited usage for some roles only.")]
@@ -26,10 +33,7 @@ namespace Orabot.Modules
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message);
-
-			var embed = CreateEmbed(message);
-			await SendQuote(Context.User, embed);
+			HandleQuote(Context.Guild, message);
 		}
 
 		[Command("quote")]
@@ -37,16 +41,13 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <messageId>`")]
 		public async Task Quote(ulong messageId)
 		{
-			if (!TryGetMessage(Context.Guild, messageId, out var message, Context.Channel as SocketTextChannel))
+			if (!_quotingService.TryGetMessage(Context.Guild, Context.Client.CurrentUser, messageId, out var message, Context.Channel as SocketTextChannel))
 			{
 				await ReplyAsync("No such message found in this server.");
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var embed = CreateEmbed(message);
-			await SendQuote(Context.User, embed);
+			HandleQuote(message);
 		}
 
 		[Command("quote")]
@@ -55,23 +56,20 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <firstMessageId> <lastMessageId>`")]
 		public async Task Quote(ulong firstMessageId, ulong lastMessageId)
 		{
-			if (!TryGetMessage(Context.Guild, firstMessageId, out var message, Context.Channel as SocketTextChannel))
+			if (!_quotingService.TryGetMessage(Context.Guild, Context.Client.CurrentUser, firstMessageId, out var message, Context.Channel as SocketTextChannel))
 			{
 				await ReplyAsync("No such message found in this server.");
 				return;
 			}
 
-			var messages = await GetMessageList(message.Channel as SocketTextChannel, firstMessageId, lastMessageId);
+			var messages = await _quotingService.GetMessageList(message.Channel as SocketTextChannel, firstMessageId, lastMessageId);
 			if (!messages.Any())
 			{
 				await ReplyAsync("No such messages found in the target channel.");
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var embed = CreateEmbed(messages);
-			await SendQuote(Context.User, embed);
+			HandleQuote(messages);
 		}
 
 		[Command("quote")]
@@ -79,7 +77,7 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <message_identifier>`")]
 		public async Task Quote(DiscordMessageIdentifier messageIdentifier)
 		{
-			if (!TryGetChannel(Context.Guild, messageIdentifier.ChannelId, out var channel))
+			if (!_quotingService.TryGetChannel(Context.Guild, messageIdentifier.ChannelId, out var channel))
 			{
 				return;
 			}
@@ -91,10 +89,7 @@ namespace Orabot.Modules
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var embed = CreateEmbed(message);
-			await SendQuote(Context.User, embed);
+			HandleQuote(message);
 		}
 
 		[Command("quote")]
@@ -102,22 +97,19 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <firstMessageIdentifier> <lastMessageIdentifier>`")]
 		public async Task Quote(DiscordMessageIdentifier firstMessageIdentifier, DiscordMessageIdentifier lastMessageIdentifier)
 		{
-			if (!TryGetChannel(Context.Guild, firstMessageIdentifier.ChannelId, out var channel))
+			if (!_quotingService.TryGetChannel(Context.Guild, firstMessageIdentifier.ChannelId, out var channel))
 			{
 				return;
 			}
 
-			var messages = await GetMessageList(channel, firstMessageIdentifier.MessageId, lastMessageIdentifier.MessageId);
+			var messages = await _quotingService.GetMessageList(channel, firstMessageIdentifier.MessageId, lastMessageIdentifier.MessageId);
 			if (!messages.Any())
 			{
 				await ReplyAsync("No such messages found in the specified channel.");
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var embed = CreateEmbed(messages);
-			await SendQuote(Context.User, embed);
+			HandleQuote(messages);
 		}
 
 		[Command("quote")]
@@ -125,17 +117,14 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <#channel_name> <messageId>`")]
 		public async Task Quote(string channelMention, ulong messageId)
 		{
-			if (!TryGetChannel(channelMention, out var channel))
+			if (!_quotingService.TryGetChannel(Context.Guild, channelMention, out var channel))
 			{
 				await ReplyAsync("Unknown channel specified.");
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
 			var message = await channel.GetMessageAsync(messageId);
-			var embed = CreateEmbed(message);
-			await SendQuote(Context.User, embed);
+			HandleQuote(message);
 		}
 
 		[Command("quote")]
@@ -144,17 +133,14 @@ namespace Orabot.Modules
 		[Remarks("Usage: `quote <#channel_name> <firstMessageId> <lastMessageId>`")]
 		public async Task Quote(string channelMention, ulong firstMessageId, ulong lastMessageId)
 		{
-			if (!TryGetChannel(channelMention, out var channel))
+			if (!_quotingService.TryGetChannel(Context.Guild, channelMention, out var channel))
 			{
 				await ReplyAsync("Unknown channel specified.");
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var messages = await GetMessageList(channel, firstMessageId, lastMessageId);
-			var embed = CreateEmbed(messages);
-			await SendQuote(Context.User, embed);
+			var messages = await _quotingService.GetMessageList(channel, firstMessageId, lastMessageId);
+			HandleQuote(messages);
 		}
 
 		[Command("quote")]
@@ -171,16 +157,13 @@ namespace Orabot.Modules
 				return;
 			}
 
-			if (!TryGetGuild(guildId, out var guild) || !TryGetChannel(guild, channelId, out var channel))
+			if (!_quotingService.TryGetGuild(Context.Client, guildId, out var guild) || !_quotingService.TryGetChannel(guild, channelId, out var channel))
 			{
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
 			var message = await channel.GetMessageAsync(messageId);
-			var embed = CreateEmbed(message);
-			await SendQuote(Context.User, embed);
+			HandleQuote(message);
 		}
 
 		[Command("quote")]
@@ -198,7 +181,7 @@ namespace Orabot.Modules
 				return;
 			}
 
-			if (!TryGetGuild(guildId, out var guild) || !TryGetChannel(guild, channelId, out var channel))
+			if (!_quotingService.TryGetGuild(Context.Client, guildId, out var guild) || !_quotingService.TryGetChannel(guild, channelId, out var channel))
 			{
 				return;
 			}
@@ -209,262 +192,31 @@ namespace Orabot.Modules
 				return;
 			}
 
-			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
-
-			var messages = await GetMessageList(channel, firstMessageId, lastMessageId);
-			var embed = CreateEmbed(messages);
-			await SendQuote(Context.User, embed);
+			var messages = await _quotingService.GetMessageList(channel, firstMessageId, lastMessageId);
+			HandleQuote(messages);
 		}
 
 		#region Private methods
 
-		private static IEnumerable<IUser> GetChannelUsers(IChannel channel)
+		private async void HandleQuote(SocketGuild guild, string message)
 		{
-			var usersAdapter = channel.GetUsersAsync();
-			var users = usersAdapter.ToList().Result.SelectMany(x => x);
-
-			return users;
+			var embed = _quotingService.CreateEmbed(guild, message);
+			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
+			await SendQuote(Context.User, embed);
 		}
 
-		private static Dictionary<string, List<IUser>> GetChannelUsersDictionary(IChannel channel)
+		private async void HandleQuote(IMessage message)
 		{
-			var users = GetChannelUsers(channel).ToList();
-			var usersByUsername = users.GroupBy(x => x.Username);
-			var usersByNickname = users.Cast<SocketGuildUser>().Where(x => !string.IsNullOrWhiteSpace(x.Nickname)).GroupBy(x => x.Nickname);
-			var usersByNameTmp = usersByNickname.Union(usersByUsername).GroupBy(x => x.Key);
-			var usersByName = usersByNameTmp.ToDictionary(x => x.Key, y => y.SelectMany(z => z).ToList());
-
-			return usersByName;
+			var embed = _quotingService.CreateEmbed(message);
+			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
+			await SendQuote(Context.User, embed);
 		}
 
-		private bool TryGetChannel(string channelMention, out SocketTextChannel channel)
+		private async void HandleQuote(IList<IMessage> messages)
 		{
-			foreach (var socketGuildChannel in Context.Guild.Channels)
-			{
-				if (socketGuildChannel is SocketTextChannel tmpChannel && tmpChannel.Mention == channelMention)
-				{
-					channel = tmpChannel;
-					return true;
-				}
-			}
-
-			channel = null;
-			return false;
-		}
-
-		private static bool TryGetChannel(SocketGuild guild, ulong channelId, out SocketTextChannel channel)
-		{
-			foreach (var socketGuildChannel in guild.Channels)
-			{
-				if (socketGuildChannel is SocketTextChannel tmpChannel && tmpChannel.Id == channelId)
-				{
-					channel = tmpChannel;
-					return true;
-				}
-			}
-
-			channel = null;
-			return false;
-		}
-
-		private bool TryGetGuild(ulong guildId, out SocketGuild guild)
-		{
-			foreach (var clientGuild in Context.Client.Guilds)
-			{
-				if (clientGuild.Id == guildId)
-				{
-					guild = clientGuild;
-					return true;
-				}
-			}
-
-			guild = null;
-			return false;
-		}
-
-		private bool TryGetMessage(SocketGuild guild, ulong messageId, out IMessage message, SocketTextChannel preferredChannel = null)
-		{
-			var channels = new List<SocketTextChannel>(guild.TextChannels);
-			if (preferredChannel != null)
-			{
-				channels.Insert(0, preferredChannel);
-			}
-
-			foreach (var channel in channels)
-			{
-				if (!channel.Users.Any(x => x.Username == Context.Client.CurrentUser.Username && x.DiscriminatorValue == Context.Client.CurrentUser.DiscriminatorValue))
-				{
-					continue;
-				}
-
-				message = channel.GetMessageAsync(messageId).Result;
-				if (message != null)
-				{
-					return true;
-				}
-			}
-
-			message = null;
-			return false;
-		}
-
-		private async Task<List<IMessage>> GetMessageList(SocketTextChannel channel, ulong firstMessageId, ulong lastMessageId)
-		{
-			var firstMessage = await channel.GetMessageAsync(firstMessageId);
-			if (firstMessage == null)
-			{
-				return new List<IMessage>();
-			}
-
-			var downloadedMessages = channel.GetMessagesAsync(firstMessageId, Direction.After, 25);
-			var firstPage = (await downloadedMessages.First()).ToArray();
-
-			var messages = new List<IMessage>
-			{
-				firstMessage
-			};
-
-			for (var i = firstPage.Length - 1; i >= 0; i--)
-			{
-				if (firstPage[i].Author == firstMessage.Author)
-				{
-					messages.Add(firstPage[i]);
-				}
-
-				if (firstPage[i].Id == lastMessageId)
-				{
-					break;
-				}
-			}
-
-			return messages;
-		}
-
-		private static bool TryGetQuoteAuthorAndTimestamp(string messageLine, IChannel channel, out IUser author, out string authorName, out string timestamp)
-		{
-			var usersByName = GetChannelUsersDictionary(channel);
-
-			foreach (var userList in usersByName)
-			{
-				if (messageLine.StartsWith(userList.Key))
-				{
-					author = userList.Value.Count == 1 ? userList.Value.First() : null;
-					authorName = userList.Key;
-
-					var trim = messageLine.Substring(userList.Key.Length);
-					timestamp = trim.Trim();
-
-					return true;
-				}
-			}
-
-			author = null;
-			authorName = null;
-			timestamp = null;
-			return false;
-		}
-
-		private static EmbedAuthorBuilder BuildAuthorEmbed(IUser author, string authorNameToUse)
-		{
-			if (author == null)
-			{
-				return null;
-			}
-
-			return new EmbedAuthorBuilder
-			{
-				Name = authorNameToUse,
-				IconUrl = author.GetAvatarUrl()
-			};
-		}
-
-		private static EmbedFooterBuilder BuildFooterEmbed(IChannel referredChannel, string timestamp)
-		{
-			var footerText = string.Empty;
-
-			if (!string.IsNullOrEmpty(timestamp))
-			{
-				footerText = timestamp;
-			}
-
-			if (referredChannel != null)
-			{
-				if (!string.IsNullOrEmpty(timestamp))
-				{
-					footerText += ", ";
-				}
-
-				footerText += $"in #{referredChannel.Name}";
-			}
-
-			return new EmbedFooterBuilder
-			{
-				Text = footerText
-			};
-		}
-
-		private Embed CreateEmbed(string message)
-		{
-			var lines = message.Split('\n').ToList();
-
-			if (TryGetChannel(lines[0].Trim(), out var referredChannel))
-			{
-				lines.RemoveAt(0);
-			}
-
-			if (TryGetQuoteAuthorAndTimestamp(lines[0].Trim(), Context.Channel, out var author, out var authorName, out var timestamp))
-			{
-				lines.RemoveAt(0);
-			}
-
-			var embed = new EmbedBuilder
-			{
-				Color = Color.Blue,
-				Author = BuildAuthorEmbed(author, authorName),
-				Description = string.Join("\n", lines),
-				Footer = BuildFooterEmbed(referredChannel, timestamp)
-			};
-
-			return embed.Build();
-		}
-
-		private static Embed CreateEmbed(IMessage message)
-		{
-			var author = message.Author;
-			var authorName = (author as SocketGuildUser)?.Nickname ?? author.Username;
-			var referredChannel = message.Channel;
-			var timestamp = message.Timestamp.ToString("s").Replace('T', ' ') + " UTC";
-			var descriptionText = $"{message.Content}\n\n[Original message]({message.GetJumpUrl()})";
-
-			var embed = new EmbedBuilder
-			{
-				Color = Color.Blue,
-				Author = BuildAuthorEmbed(author, authorName),
-				Description = descriptionText,
-				Footer = BuildFooterEmbed(referredChannel, timestamp)
-			};
-
-			return embed.Build();
-		}
-
-		private static Embed CreateEmbed(IList<IMessage> messages)
-		{
-			var message = messages.First();
-			var author = message.Author;
-			var authorName = (author as SocketGuildUser)?.Nickname ?? author.Username;
-			var referredChannel = message.Channel;
-			var timestamp = message.Timestamp.ToString("s").Replace('T', ' ') + " UTC";
-			var descriptionText = $"{string.Join("\n", messages.Select(x => x.Content))}\n\n[Original message]({message.GetJumpUrl()})";
-
-			var embed = new EmbedBuilder
-			{
-				Color = Color.Blue,
-				Author = BuildAuthorEmbed(author, authorName),
-				Description = descriptionText,
-				Footer = BuildFooterEmbed(referredChannel, timestamp)
-			};
-
-			return embed.Build();
+			var embed = _quotingService.CreateEmbed(messages);
+			await Context.Channel.DeleteMessageAsync(Context.Message, RequestOptions.Default);
+			await SendQuote(Context.User, embed);
 		}
 
 		private async Task SendQuote(IUser user, Embed embed)
